Make UI fully usable using a controller
This commit focuses on making the UI completely usable using a controller so that a user won't have to switch between touch and the controller constantly. There is an exception to this that is `LogActivity`, it cannot be navigated through using a controller (but can be exited using the B button) and this isn't going to be fixed as we plan to replace it eventually anyway.